+++ include/render/vertexBuffer.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstddef>
+
 #include <utils/glew.hpp>
 
 namespace fcm {
@@ -8,7 +10,7 @@ struct VertexBuffer
 {
     glew::GLHandle handle;
 
-    VertexBuffer(const void* data, unsigned int size);
+    VertexBuffer(const void* data, size_t size);
     ~VertexBuffer();
 
     void bind() const;
